Closes #89 getting names for rats and clients or ids for rats saves ids for the current session.
